refactor: extract archive staging helpers
(cherry picked from commit 6565ae1857b1db906b9c429ec651b9d5f1eb0ea8)
